[rollout] refactor: integrate Qwen-Image step execution
