+++ verl_omni/__init__.py
@@ -30,7 +30,6 @@ except Exception:
 
 # Import pipelines / rollout / reward loop / engines to auto-register them
 # Apply model patches and auto-register pipelines / rollout / reward loop / engines
-import verl_omni.experimental  # noqa: E402, F401
 import verl_omni.models  # noqa: E402, F401
 import verl_omni.pipelines  # noqa: E402, F401
 import verl_omni.reward_loop  # noqa: E402, F401
